+++ src/gmt_mgg_header2.c
@@ -1,4 +1,4 @@
-/*	$Id: gmt_mgg_header2.c,v 1.41 2010-03-22 18:55:44 guru Exp $
+/*	$Id: gmt_mgg_header2.c,v 1.42 2010-03-23 00:03:52 jluis Exp $
  *
  *	Code donated by David Divens, NOAA/NGDC
  *	Distributed under the GNU Public License (see LICENSE.TXT for details)
@@ -487,8 +487,8 @@ GMT_LONG mgg2_read_grd (struct GRD_HEADER *header, float *grid, double w, double
 		}
 	}
 	
-	header->nx = width_in;
-	header->ny = height_in;
+	header->nx = (int)width_in;
+	header->ny = (int)height_in;
 	header->x_min = w;
 	header->x_max = e;
 	header->y_min = s;
